% Generated by roxygen2 (4.0.0): do not edit by hand
\name{block_grid}
\alias{block_grid}
\title{Creates a block grid of a given data type}
\usage{
block_grid(nrow, ncol = nrow, type = "data.frame", fill = get_fill(""))
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{type}{data type of object}

\item{fill}{default fill color}
}
\description{
Creates a block grid of a given data type
}
\examples{
grid1 = block_grid(10, 10)
grid1 = block_grid(10, 10, 'matrix')
grid1[1]  = 'red'
grid1
grid1 = block_grid(10, type = 'vector')
grid1[1]  = 'red'
grid1
}

